Use KoboldCPP SSE streaming endpoint
The code to interact with the /extra/generate/stream endpoint of KoboldCPP API
is basically the same as for OpenAI streaming mode. The difference is the API path
and the fact that KoboldCPP API sends also other types of lines, not just "data:" so we
can ignore them. The path of the token field in the returned JSON is also different than OpenAI.
Other than that there should be no differences.

For reference here is the KoboldCPP side:
https://github.com/LostRuins/koboldcpp/blob/concedo/koboldcpp.py#L593